git-svn-id: https://svn.r-project.org/R/branches/djm-source@39697 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/duplicate.c
@@ -87,7 +87,7 @@
 static SEXP duplicate1(SEXP);
 
 #ifdef R_PROFILING
-static unsigned long duplicate_counter = -1;
+static unsigned long duplicate_counter = (unsigned long)-1;
 
 unsigned long  attribute_hidden
 get_duplicate_counter(void)
